+++ cmd/hydra/extras.go
@@ -84,7 +84,9 @@ func startExtraEngines(ctx context.Context, cfg *config.HydraConfig, engineCfgs 
 			}
 		}
 
-		ann := engine.NewHoardAnnouncer(proc.Client(), engine.DefaultSingleBinding(le.cfg.ListenPort, le.cfg.EnableIPv6, "hoard", le.cfg.AnnounceRateLimit))
+		ann := engine.NewHoardAnnouncer(proc.Client(), engine.ApplyAnnounceEgress(
+			engine.DefaultSingleBinding(le.cfg.ListenPort, le.cfg.EnableIPv6, "hoard", le.cfg.AnnounceRateLimit),
+			le.cfg.AnnounceProxy, le.cfg.AnnounceIP, le.cfg.Socks5OutboundHost, "hoard"))
 		if le.hoard != nil {
 			ann.OnObservation = le.hoard.ObserveAnnounce
 			le.hoard.SetBootstrapAnnounce(ann.BootstrapAnnounce)